+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -19,7 +19,7 @@
 -->
 ## Upcoming Release
 
-* Get and update global parameters in Data Factory.
+* Added global parameters to Data Factory.
 
 ## Version 1.8.2
 * Updated ADF .Net SDK version to 4.9.0